feat: streaming flow validation
