+++ core/api-server/tests/pipelines-store.js
@@ -408,72 +408,6 @@ describe('Store/Pipelines', () => {
             expect(response.body.options.batchTolerance).to.equal(80);
             expect(response.body.options.progressVerbosityLevel).to.equal('info');
         });
-        it('should succeed to store flows streaming pipeline', async () => {
-            const pipeline = {
-                name: 'streaming-flow',
-                experimentName: 'main',
-                kind: 'stream',
-                nodes: [
-                    {
-                        nodeName: 'A',
-                        algorithmName: 'green-alg',
-                        input: [],
-                        stateType: 'stateful',
-                        kind: 'algorithm'
-                    },
-                    {
-                        nodeName: 'B',
-                        algorithmName: 'green-alg',
-                        input: [],
-                        stateType: 'stateless',
-                        kind: 'algorithm'
-                    },
-                    {
-                        nodeName: 'C',
-                        algorithmName: 'green-alg',
-                        input: [],
-                        stateType: 'stateless',
-                        kind: 'algorithm'
-                    },
-                    {
-                        nodeName: 'D',
-                        algorithmName: 'green-alg',
-                        input: [],
-                        stateType: 'stateless',
-                        kind: 'algorithm'
-                    },
-                    {
-                        nodeName: 'E',
-                        algorithmName: 'green-alg',
-                        input: [],
-                        stateType: 'stateless',
-                        kind: 'algorithm'
-                    }
-                ],
-                streaming: {
-                    flows: {
-                        analyze0: 'A >> B >> C >> D >> B >> A',
-                        analyze1: 'A >> B&C , C >> D',
-                        analyze2: 'A >> B&C >> D',
-                        analyze3: 'A >> B >> C >> D >> A',
-                        analyze4: 'A >> B&C&D >> E'
-                    }
-                },
-                priority: 3,
-                options: {
-                    batchTolerance: 80,
-                    progressVerbosityLevel: 'info',
-                    ttl: 3600
-                }
-            };
-            const options = {
-                uri: restPath,
-                body: pipeline
-            };
-            const response = await request(options);
-            expect(response.response.statusCode).to.equal(HttpStatus.CREATED);
-            expect(response.body).to.deep.equal(pipeline);
-        });
         it('should succeed to store pip', async () => {
             const pipeline = {
                 name: 'pipeline_in_pipeline',
